+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* [improvement] `shelly add` should ask about databases instead of database
+
 ## 0.4.17 / 2013-12-03
 
 * [bugfix] Remove ssh_key from Shelly Cloud only if it was already uploaded